fix: mobile view fixes
