+++ frontend/src/components/MobileLayout.vue
@@ -50,18 +50,16 @@ export default {
           isActive: /MyTasks|Task/g.test(this.$route.name),
         },
         {
-          name: 'Teams',
+          name: 'Spaces',
           icon: LucideLayoutGrid,
-          route: { name: 'Teams' },
+          route: { name: 'Spaces' },
           isActive: [
-            'Teams',
-            'TeamOverview',
-            'ProjectOverview',
-            'ProjectDiscussions',
-            'ProjectDiscussion',
-            'ProjectDiscussionNew',
-            'ProjectTasks',
-            'ProjectTaskDetail',
+            'Spaces',
+            'Space',
+            'SpaceDiscussions',
+            'SpaceDiscussion',
+            'SpaceTasks',
+            'SpaceTask',
           ].includes(this.$route.name),
         },
         {
